Fix misplaced, out-of-date test.
